requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.32.0
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.4.0
django-celery-results==2.3.1
django-ckeditor==6.4.0
django-cloudinary-storage==0.3.0
django-ipware==4.0.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-solo==2.0.0
django-timezone-field==5.0
idna==3.3
ipapi==1.0.4
kombu==5.2.4
packaging==21.3
Pillow==9.4.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.5
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
